move Set to set.py and Dict to dict.py
